requirements.txt
kaggle
pandas
numpy
apache-airflow-providers-postgres
seaborn
matplotlib
seaborn
